Day 10 Python and Rust solutions
Rust takes noticeably longer than other days at ~150ms on a release
build. I think that much faster solutions exist if you make some
simplifying assumptions about your input (e.g. fast exponentiation
solution with modulo-2 counter combinations being reachable in some
combination of single presses of the existing buttons), but for now I
did not want to make such assumptions, and resolved to use Z3, which
makes it run a bit slower.

Ideally if I want to get this runtime down in line with the other
days in the future, I should try such special cases first with a limit,
and then fallback to Z3. This will allow fast solutions for real inputs
but still achieve the desired general solution. But that is not done for
now.